Trying out Arrays??
By looking at the snake game, I’m trying to make an array that will
“split” the egg in half, or for now move the lefts side half of the egg
to the left, off the screen. I don’t have it working yet. I finished
most of the basic animations and how they’ll look. Figuring out the
arrays will be the hardest part of the project so that’s why I’m trying
to do it now.