Moving on...
Decided not to continue working on the array. I got it to work, but
there were some weird bugs with the animation so i decided to comment
it out for now and maybe go back to it later. Right now I’m putting
together the animations, and I might use the array for something
simpler that will hopefully be less buggy (just the chicken moving off
the screen)